data seeding non-prod only

+++ src/todo/todo.service.ts
@@ -3,10 +3,14 @@ import { CreateTodoDto } from './dto/create-todo.dto';
 import { UpdateTodoDto } from './dto/update-todo.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
 import todos from './fixtures/todos';
+import { EnvService } from 'src/env/env.service';
 
 @Injectable()
 export class TodoService {
-  constructor(private readonly prismaService: PrismaService) {}
+  constructor(
+    private readonly envService: EnvService,
+    private readonly prismaService: PrismaService,
+  ) {}
 
   create(createTodoDto: CreateTodoDto) {
     return 'This action adds a new todo';
@@ -28,15 +32,19 @@ export class TodoService {
     return `This action removes a #${id} todo`;
   }
 
+  async clear(): Promise<void> {
+    if (this.envService.isProd) {
+      throw new Error('Data seeding is not allowed in production mode.');
+    }
+    await this.prismaService.todo.deleteMany({});
+  }
+
   async seed(): Promise<void> {
-    // TODO: non-prod only
+    if (this.envService.isProd) {
+      throw new Error('Data seeding is not allowed in production mode.');
+    }
     await this.prismaService.todo.createMany({
       data: todos,
     });
-  }
-
-  async clear(): Promise<void> {
-    // TODO: non-prod only
-    await this.prismaService.todo.deleteMany({});
   }
 }
